restructuring + refactor
