+++ src/components/BtcFearAndGreed.js
