+++ src/components/CandlesChart.js
