+++ src/components/CoinModal.js
