+++ src/services/mockdata/fearAndGreed.js
