+++ src/services/mockdata/index.js
@@ -1,0 +1,22 @@
+import { news } from './news';
+import { newsAlternative } from './newsAlternative';
+import { coins } from './coins';
+import {trending} from './trending'
+import { globaldata } from './globaldata';
+import {exchanges} from './exchanges'
+import { fearAndGreed } from './fearAndGreed';
+import { btc } from './btc';
+import { youtube } from './youtube';
+
+// MOCK API's Data
+export const mockData = {
+      cryptoNewsApi : news,
+      cryptoNewsApiAlt: newsAlternative,
+      coins: coins,
+      btc: btc,
+      trending: trending,
+      fearAndGreed: fearAndGreed,
+      globalData: globaldata,
+      exchanges: exchanges,
+      youtube: youtube
+};
