--- src/utils/Endpoints.js
@@ -1,17 +1,0 @@
-export const endpoints  = [
-    'https://crypto-news-api.b4a.app/news/theblockcrypto',
-    'https://api.coingecko.com/api/v3/simple/price?ids=bitcoin&vs_currencies=usd',
-    'https://api.coingecko.com/api/v3/coins/markets?vs_currency=usd&order=market_cap_desc&per_page=50&page=1&sparkline=false',
-    'https://api.coingecko.com/api/v3/search/trending',
-    'https://api.alternative.me/fng/?limit=14',
-    'https://api.coingecko.com/api/v3/global',
-    'https://www.googleapis.com/youtube/v3/search?key=AIzaSyB_DuvsQ8DclIA0-RihVYDzCNwjd0fQHes&channelId=UCqK_GSMbpiV8spgD3ZGloSw&part=snippet&maxResults=1&order=date&type=video',
-    'https://www.googleapis.com/youtube/v3/search?key=AIzaSyB_DuvsQ8DclIA0-RihVYDzCNwjd0fQHes&channelId=UCCatR7nWbYrkVXdxXb4cGXw&part=snippet&maxResults=3&order=date&type=video',
-    'https://www.googleapis.com/youtube/v3/search?key=AIzaSyB_DuvsQ8DclIA0-RihVYDzCNwjd0fQHes&channelId=UCwGflGmzevf4fcm-z8E-twA&part=snippet&maxResults=3&order=date&type=video'
-]
-
-export default {
-    endpoints,
-}
-
-//https://newsdata.io/api/1/news?apikey=pub_167632e1074e9c30ea0b34c3119c202cf3bd5&q=cryptocurrency%20,%20crypto%20,%20blockchain%20,%20bitcoin&language=en
