--- src/utils/RoutesList.js
@@ -1,27 +1,0 @@
-import Exchanges from "../views/Exchanges";
-import AboutUs from "../views/AboutUs";
-import NewsLetter from "../views/NewsLetter";
-import PrivacyPolicy from "../views/PrivacyPolicy";
-import NewsArticle from "../views/NewsArticle";
-import News from "../views/News";
-import Chart from "../views/Chart";
-import Resources from "../views/Resources";
-import Home from "../views/Home";
-import Register from "../components/Register";
-import Login from "../components/Login";
-import UserPage from "../components/UserPage";
-
-export {
-  AboutUs,
-  NewsLetter,
-  Exchanges,
-  PrivacyPolicy,
-  NewsArticle,
-  News,
-  Chart,
-  Resources,
-  Home,
-  Register,
-  Login,
-  UserPage,
-};
