+++ src/views/PrivacyPolicy.js
