+++ src/views/index.js
@@ -1,0 +1,27 @@
+import Exchanges from "./Exchanges";
+import AboutUs from "./AboutUs";
+import NewsLetter from "./NewsLetter";
+import PrivacyPolicy from "./PrivacyPolicy";
+import NewsArticle from "./NewsArticle";
+import News from "./News";
+import Chart from "./Chart";
+import Resources from "./Resources";
+import Home from "./Home";
+import Register from "../components/Register";
+import Login from "../components/Login";
+import UserPage from "../components/UserPage";
+
+export {
+  AboutUs,
+  NewsLetter,
+  Exchanges,
+  PrivacyPolicy,
+  NewsArticle,
+  News,
+  Chart,
+  Resources,
+  Home,
+  Register,
+  Login,
+  UserPage,
+};
